Pre-norminetted version, better error reporting
