+++ maps/invalid.rt
@@ -1,5 +1,5 @@
 pl 1,.3,3 0,0.5,-1 233,233,233
-A 0000.9999 5,5,70
+A 0000.9999 55,5,70
 
 C  8000,0,99.00005 0.004,0.2,-1.0 70
 L -4000,0,30 .17 255,255,255
@@ -9,4 +9,3 @@ sp 0,0,20.0000 20 255,255,0
 cy 50.0,0.0,20.6 0,0,1.0 14.2 21.42 10,0,255
 cy -6.0,20,0. 0,-1.0,0.4 1 012 10,10,12
 sp 0,1.0,1 045 01,25,23
-
