--- src/load_data_utils.c
@@ -1,22 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   load_data_utils.c                                  :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: plouda <marvin@42.fr>                      +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/17 14:03:53 by plouda            #+#    #+#             */
-/*   Updated: 2023/11/17 14:28:06 by plouda           ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../header/minirt.h"
-
-/* void	get_rgb(t_rt *rt, void *object, char *id)
-{
-	if (!ft_strncmp(id, "A", 1))
-		object = (t_ambient *)object;
-	rt->object->rgb[R] = ft_atoi(rgb[0]);
-	rt->object->rgb[G] = ft_atoi(rgb[1]);
-	rt->object->rgb[B] = ft_atoi(rgb[2]);
-} */
